일부 파일 이름 변경 및 엔터제거, 미사용 import 제거
- 역할에 맞게 js 파일 이름 변경
 - room -> index
 - game -> room
