+++ src/main/java/com/example/quiz/config/SecurityConfig.java
@@ -4,7 +4,6 @@ package com.example.quiz.config;
 import com.example.quiz.config.auth.HttpCookieOAuth2AuthorizationRequestRepository;
 import com.example.quiz.entity.user.CustomAuthenticationSuccessHandler;
 import com.example.quiz.jwt.JWTRequestFilter;
-import com.example.quiz.jwt.JwtUtil;
 import com.example.quiz.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
