+++ src/main/java/com/example/quiz/config/auth/CookieUtils.java
@@ -1,6 +1,5 @@
 package com.example.quiz.config.auth;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
