+++ src/main/resources/templates/quiz.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>Quiz In-Game</title>
     <!-- SockJS, stomp.js -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.4.0/sockjs.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.5.0/sockjs.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"></script>
     <!-- quiz.js 연결 (JS 로직) -->
     <script defer src="/script/quiz.js"></script>
